chore: merge post-process into cli
