--- src/post-process/readme.md
@@ -1,10 +1,0 @@
-# Post Processing
-
-This script relies on a `enhance.gdb` table that contains house, senate, parcel, and census data with the names described in the `IDENTITY` object.
-
-## Environment
-
-- `conda creation --name cloud-geocoding python=3.7`
-- `activate cloud-geocoding`
-- `conda install -c esri arcpy`
-- `pip install -r requirements.txt`
