--- src/post-process/requirements.txt
@@ -1,1 +1,0 @@
-docopt
